requirements.txt
antlr4-python3-runtime==4.9.3 ; python_version >= "3.9" and python_version < "3.12"
cloudpickle==2.2.1 ; python_version >= "3.9" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "3.12" and platform_system == "Windows" or python_version >= "3.9" and python_version < "3.12" and sys_platform == "win32"
contourpy==1.0.7 ; python_version >= "3.9" and python_version < "3.12"
cvxopt==1.3.1 ; python_version >= "3.9" and python_version < "3.12"
cvxpy==1.3.1 ; python_version >= "3.9" and python_version < "3.12"
cycler==0.11.0 ; python_version >= "3.9" and python_version < "3.12"
ecos==2.0.12 ; python_version >= "3.9" and python_version < "3.12"
exceptiongroup==1.1.1 ; python_version >= "3.9" and python_version < "3.11"
filelock==3.12.0 ; python_version >= "3.9" and python_version < "3.12"
fonttools==4.39.4 ; python_version >= "3.9" and python_version < "3.12"
gurobipy==10.0.0 ; python_version >= "3.9" and python_version < "3.12"
gym==0.21.0 ; python_version >= "3.9" and python_version < "3.12"
hydra-core==1.3.2 ; python_version >= "3.9" and python_version < "3.12"
importlib-metadata==4.13.0 ; python_version >= "3.9" and python_version < "3.12"
importlib-resources==5.12.0 ; python_version >= "3.9" and python_version < "3.10"
iniconfig==2.0.0 ; python_version >= "3.9" and python_version < "3.12"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "3.12"
kiwisolver==1.4.4 ; python_version >= "3.9" and python_version < "3.12"
markupsafe==2.1.2 ; python_version >= "3.9" and python_version < "3.12"
matplotlib==3.7.1 ; python_version >= "3.9" and python_version < "3.12"
mpmath==1.3.0 ; python_version >= "3.9" and python_version < "3.12"
networkx==3.1 ; python_version >= "3.9" and python_version < "3.12"
numpy==1.24.3 ; python_version < "3.12" and python_version >= "3.9"
omegaconf==2.3.0 ; python_version >= "3.9" and python_version < "3.12"
osqp==0.6.2.post9 ; python_version >= "3.9" and python_version < "3.12"
packaging==23.1 ; python_version >= "3.9" and python_version < "3.12"
pandas==2.0.1 ; python_version >= "3.9" and python_version < "3.12"
pathlib==1.0.1 ; python_version >= "3.9" and python_version < "3.12"
pillow==9.5.0 ; python_version >= "3.9" and python_version < "3.12"
pluggy==1.0.0 ; python_version >= "3.9" and python_version < "3.12"
pyparsing==3.0.9 ; python_version >= "3.9" and python_version < "3.12"
pytest==7.3.1 ; python_version >= "3.9" and python_version < "3.12"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "3.12"
pytz==2023.3 ; python_version >= "3.9" and python_version < "3.12"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "3.12"
qdldl==0.1.7 ; python_version >= "3.9" and python_version < "3.12"
scipy==1.10.1 ; python_version >= "3.9" and python_version < "3.12"
scs==3.2.3 ; python_version >= "3.9" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.9" and python_version < "3.12"
stable-baselines3==1.8.0 ; python_version >= "3.9" and python_version < "3.12"
sympy==1.12 ; python_version >= "3.9" and python_version < "3.12"
tomli==2.0.1 ; python_version >= "3.9" and python_version < "3.11"
tqdm==4.65.0 ; python_version >= "3.9" and python_version < "3.12"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "3.12"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "3.12"
zipp==3.15.0 ; python_version >= "3.9" and python_version < "3.12"
